refactor: moves 2 functions away from services package due to not being used by any external app - changes context in UserData view to keep previous behaviour
