+++ uchileedxlogin/services/utils.py
@@ -1,31 +1,9 @@
 # Python Standard Libraries
 import logging
-from itertools import cycle
 
 
 logger = logging.getLogger(__name__)
 
-
-def validate_rut(rut):
-    """
-    Verify if the rut is valid.
-    """
-    rut = rut.upper()
-    rut = rut.replace("-", "")
-    rut = rut.replace(".", "")
-    rut = rut.strip()
-    aux = rut[:-1]
-    dv = rut[-1:]
-    revertido = list(map(int, reversed(str(aux))))
-    factors = cycle(list(range(2, 8)))
-    s = sum(d * f for d, f in zip(revertido, factors))
-    res = (-s) % 11
-    if str(res) == dv:
-        return True
-    elif dv == "K" and res == 10:
-        return True
-    else:
-        return False
 
 def get_document_type(doc_id):
     """
