+++ uchileedxlogin/users.py
@@ -67,3 +67,16 @@ def create_edxloginuser(user, have_sso, doc_id):
     except Exception as e:
         logger.error(f"create_edxloginuser failed for user: {user}, have_sso: {have_sso} and doc_id: {doc_id}, with error: {e}")
         raise Exception(f"Failed to create EdxLoginUser object for user: {user}, have_sso: {have_sso} and doc_id: {doc_id}")
+
+
+# Model permissions
+def check_permission_instructor_staff(user):
+    """
+    Check if the user has the permission uchile_instructor_staff.
+    """
+    if not user.is_anonymous:
+        if user.has_perm('uchileedxlogin.uchile_instructor_staff') or user.is_staff:
+            return True
+        logger.error(f"Insufficient permissions, user: {user} isn't staff nor has special permission.")
+    logger.error(f"Insufficient permissions, user: {user} is anonymous.")
+    return False
